+++ streams/voxel_stream_script.cpp
@@ -16,7 +16,7 @@ void VoxelStreamScript::load_voxel_block(VoxelStream::VoxelQueryData &query_data
 
 	int res;
 #if defined(ZN_GODOT)
-	if (GDVIRTUAL_CALL(_load_voxel_block, buffer_wrapper, query_data.origin_in_voxels, query_data.lod_index, res)) {
+	if (GDVIRTUAL_CALL(_load_voxel_block, buffer_wrapper, query_data.position_in_blocks, query_data.lod_index, res)) {
 		// Check if the return enum is valid
 		ERR_FAIL_INDEX(res, _RESULT_COUNT);
 		// If the block was found, grab its data from the script-facing object to our internal buffer
@@ -38,7 +38,7 @@ void VoxelStreamScript::save_voxel_block(VoxelStream::VoxelQueryData &query_data
 	buffer_wrapper.instantiate();
 	query_data.voxel_buffer.duplicate_to(buffer_wrapper->get_buffer(), true);
 #if defined(ZN_GODOT)
-	if (!GDVIRTUAL_CALL(_save_voxel_block, buffer_wrapper, query_data.origin_in_voxels, query_data.lod_index)) {
+	if (!GDVIRTUAL_CALL(_save_voxel_block, buffer_wrapper, query_data.position_in_blocks, query_data.lod_index)) {
 		WARN_PRINT_ONCE("VoxelStreamScript::_save_voxel_block is unimplemented!");
 	}
 #else
@@ -61,8 +61,8 @@ int VoxelStreamScript::get_used_channels_mask() const {
 void VoxelStreamScript::_bind_methods() {
 #if defined(ZN_GODOT)
 	// TODO Test if GDVIRTUAL can print errors properly when GDScript fails inside a different thread.
-	GDVIRTUAL_BIND(_load_voxel_block, "out_buffer", "origin_in_voxels", "lod");
-	GDVIRTUAL_BIND(_save_voxel_block, "buffer", "origin_in_voxels", "lod");
+	GDVIRTUAL_BIND(_load_voxel_block, "out_buffer", "position_in_blocks", "lod");
+	GDVIRTUAL_BIND(_save_voxel_block, "buffer", "position_in_blocks", "lod");
 	GDVIRTUAL_BIND(_get_used_channels_mask);
 #endif